Post display page, including the idea of Comments
This wires up the post show page. It's a bit meager right now, but I
wanted to take a crack at the most interesting part -- the comment
system.

To this end I have found an interesting library 'awesome_nested_set'
which is way more interesting than just setting a parent_id on the
comments to refer to their parents. You still have to set a parent_id,
but it uses the IDs of comments in a clever way so it improves the
efficiency of some operations. Namely
-- you can select all descendants of a comment: one SQL query
-- deleting a comment and its descendants: one query
-- more benefits here:
  http://threebit.net/tutorials/nestedset/tutorial1.html
(The downside is that the SQL for inserting a comment is pretty
ridiculous. Not a lot of queries -- but one very complicated query.
Fortunately, we can assume it's well tested).